requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.0
nltk==3.8.1
jinja2==3.1.2
python-multipart==0.0.6
scikit-learn==1.4.0
pytest==7.4.4
streamlit==1.33.0